Added Gun & Materials
I added a gun to the scene which I made VR moveable. I placed it on the podium where the user will be standing.

I also added some materials to the scene in order to just make the general scene look better. I also experimented with the material of the gun, trying to make it metallic, but it doesn't really show up as well as I hoped.