Demo 5 FINISHED!!!!
FINALLY DID IT!! Figured out how to make the balls shoot in the direction of the raycast, while also changing shooting velocity depending on how far the targets were.

I also added a lineRenderer to my gun that would appear only when the raycast was colliding with an object. This was in order to help a user aim as I was having trouble getting my triggerPoint to align with my object in space.