Set up DemoScene6 & Added Canvas
I've decided that I want to kind of build upon my demos, so I started by making a copy of demo 5, and labeling it as demo 6.

I've also added a basic canvas with a slider, button, dropdown menu, etc. This was accompanied by a world-mouse laser.

I'm committing this so that I have a place to start from scratch in case I feel the need to start over.