Added Button Functionality
The GUI now has an interactive button.

The button changes the color of the bullets being shot on click. It rotates between black, gray, and gold.

The name of the label on the button also changes on click, with the word on the button reflecting the color setting it is on.