DEMO 7 COMPLETE!!!
I added the simultaneous coroutines to all 5 targets, and I had them move at slightly different speeds to show that they are not all on the same coroutine.

A little quirk though, I can't shoot the object that the script is on, or the rest stop moving. I just have to shoot that one last!